requirements.txt
fastapi
uvicorn
sqlalchemy
pydub
requests
streamlit
openai
python-dotenv
audio-recorder-streamlit